Only attach editor to DOM if script is activated
This change was made because OmniSharp server is only started once a script is activated. When the editor is attached for all open scripts, Monaco editor will execute semantic highlighting, however the OmniSharp server wasn't started yet. Monaco will not know to redo the call to get semantic highlighting once OmniSharp is started. This way, Monaco will initialize only when the script is activated for the first time.

This also has the added benefit of a slight perf bonus of not initializing Monaco for auto-saved scripts that are restored that aren't activated yet