Consolidate all NuGet package installations in the same directory and show status message when package is being installed.
Before this change, when installing a nuget package, NuGet would download the packages in the global nuget folder (~/.nuget/packages on Linux and OSX) and NetPad would extract the downloaded package to the configured package cache directory.

This meant we would have 2 installations on disk of the same package. This change consolidates this so we only have one installation on disk, and in the package cache directory that the user configures.