Fixes an issue where EF Core does not scaffold ToTable() mappings for some tables
EF Core not mapping ToTable() mappings causes an error when we rename the DbSet property names generated on the DbContext. It complains about not being able to find the SQL table.

It seems if EF Core determines that the table name in SQL is the same as the generated DbSet property name, then it omits the ToTable() mapping.