O2HTML: use reference comparison to check for previously serialized objects
This is important in the case of C# records. By default their equality is dependant on the value of their properties, not on reference equality. Value equality results in identifying 2 distinct record-type objects as previously serialized if they have the same property values, where in reality they are different objects and we want to serialize each without considering it a cyclic reference.