Buffer raw process output so it doesn't overwhelm IPC
When testing stack overflows from infinite recursion, the process that crashed with a stack overflow outputted thousands of lines of error data. Each error was output as a separate line and came across as a separate process output that would be pushed independently over IPC, resulting in thousands of IPC calls in rapid succession. This caused the application to hang in extreme situation. Buffering this output makes a world of difference in performance and does not make the app crash.