Move window control features out of the backend and embed it directly into the Electron main process in the JS app
The reason for this change was mostly due to feeling that features like maximizing the window or opening dev tools feel out of place being on the backend, and the awkwardness of having to call the API to maximize a window. This is UI specific work and can, and should, be done entirely within the JS app without needing to involve the API app. This also makes for better performance and less network traffic.

The Electron main process code was added using the "electronize add hosthook" command which added the "ElectronHostHook" directory. This also now gives us a foundation to add more main process functionality directly in the JS app without needing to involve the backend.