Fixes race condition bug with OmniSharp and code buffer update requests
OmniSharp calls can come back with stale or invalid data because a script code buffer update is still in progress and hasn't completed yet. This change ensures that the SPA app waits on OmniSharp calls until code buffer updates are complete.

This fix is the permanent fix to the temp fix put in place in b55272ef and replaces some of that commit's code.