Fixes regression caused by 8b6927ac
Script fails to execute with System.MissingMethodException: Entry point not found in assembly NetPadScript when a data connection is assigned to script.

The issue is caused by the compiled data connection assembly having the same name as the compiled script assembly, so the data connection assembly overwrites the script assembly causing the issue.